Reorganization of database representation types
I moved the firebase table representation types into a firebase/types folder so that they're all in the same location and can easily be edited as well as integrate with firebase a lot cleaner
