+++ .idea/misc.xml
@@ -9,6 +9,7 @@
         <entry key="..\:/Users/abiiy/AndroidStudioProjects/GeoSnapApp/app/src/main/res/layout/activity_register_spinner0.xml" value="0.3536458333333333" />
         <entry key="..\:/Users/abiiy/AndroidStudioProjects/GeoSnapApp/app/src/main/res/layout/activity_register_spinner1.xml" value="0.3536458333333333" />
         <entry key="..\:/Users/abiiy/AndroidStudioProjects/GeoSnapApp/app/src/main/res/layout/activity_register_spinner2.xml" value="0.18385416666666668" />
+        <entry key="..\:/Users/abiiy/AndroidStudioProjects/GeoSnapApp/app/src/main/res/layout/fragment_add.xml" value="0.33125" />
         <entry key="..\:/Users/diann/StudioProjects/GeoSnapApp/app/src/main/res/layout/activity_create_station.xml" value="0.1" />
         <entry key="..\:/Users/diann/StudioProjects/GeoSnapApp/app/src/main/res/layout/activity_expand.xml" value="0.1" />
         <entry key="..\:/Users/diann/StudioProjects/GeoSnapApp/app/src/main/res/layout/activity_listfragment_recycler.xml" value="0.1" />
