+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
             android:label="@string/title_activity_register__org__admin"
             android:theme="@style/AppTheme.NoActionBar"  android:hasCode="true"></activity>
         <activity android:name="org.kashmirworldfoundation.WildlifeGeoSnap.StationListActivity" />
-        <activity android:name="org.kashmirworldfoundation.WildlifeGeoSnap.CreateStudy" />
+        <activity android:name="org.kashmirworldfoundation.WildlifeGeoSnap.study.CreateStudyActivity" />
         <activity android:name="org.kashmirworldfoundation.WildlifeGeoSnap.GoogleMapActivity" /> <!-- <activity android:name=".Expand2" /> -->
         <activity
             android:name="org.kashmirworldfoundation.WildlifeGeoSnap.ExpandActivity"
