+++ app/src/main/java/org/kashmirworldfoundation/WildlifeGeoSnap/AddWildlifeSightingActivity.java
@@ -30,6 +30,9 @@ import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 import com.google.firebase.storage.UploadTask;
 
+import org.kashmirworldfoundation.WildlifeGeoSnap.firebase.types.Member;
+import org.kashmirworldfoundation.WildlifeGeoSnap.firebase.types.WildlifeSighting;
+
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
