+++ app/src/main/java/org/kashmirworldfoundation/WildlifeGeoSnap/RegisterSpinner2.java
@@ -19,6 +19,7 @@ import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 
 import org.kashmirworldfoundation.WildlifeGeoSnap.auth.user.register.RegisterActivity;
+import org.kashmirworldfoundation.WildlifeGeoSnap.firebase.types.Org;
 
 import java.util.ArrayList;
 import java.util.LinkedHashSet;
