+++ app/src/main/java/org/kashmirworldfoundation/WildlifeGeoSnap/firebase/types/Marker.java
@@ -1,4 +1,4 @@
-package org.kashmirworldfoundation.WildlifeGeoSnap;
+package org.kashmirworldfoundation.WildlifeGeoSnap.firebase.types;
 
 public class Marker {
     private String author;
