+++ app/src/main/java/org/kashmirworldfoundation/WildlifeGeoSnap/firebase/types/Member.java
@@ -1,4 +1,4 @@
-package org.kashmirworldfoundation.WildlifeGeoSnap;
+package org.kashmirworldfoundation.WildlifeGeoSnap.firebase.types;
