+++ app/src/main/java/org/kashmirworldfoundation/WildlifeGeoSnap/firebase/types/Org.java
@@ -1,4 +1,4 @@
-package org.kashmirworldfoundation.WildlifeGeoSnap;
+package org.kashmirworldfoundation.WildlifeGeoSnap.firebase.types;
 
 public class Org {
     private  String OrgEmail;
