+++ app/src/main/java/org/kashmirworldfoundation/WildlifeGeoSnap/firebase/types/Study.java
@@ -1,4 +1,4 @@
-package org.kashmirworldfoundation.WildlifeGeoSnap;
+package org.kashmirworldfoundation.WildlifeGeoSnap.firebase.types;
 
 import android.os.Parcel;
 import android.os.Parcelable;
