+++ app/src/main/res/layout/activity_create_study.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context="org.kashmirworldfoundation.WildlifeGeoSnap.CreateStudy">
+    tools:context="org.kashmirworldfoundation.WildlifeGeoSnap.study.CreateStudyActivity">
 
     " />
